Correlations - Computing Observables and Other Analyses

fermions.GenCorrelation — Method

GenCorrelation(state, operator)

Calculate the expectation value ⟨state|operator|state⟩.

Examples

julia> state = Dict(Bool.([1, 0]) => 0.5, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 0.5
  [0, 1] => -0.5

julia> operator = [("+-", [1, 2], 1.0)]
1-element Vector{Tuple{String, Vector{Int64}, Float64}}:
 ("+-", [1, 2], 1.0)

julia> GenCorrelation(state, operator)
-0.5

fermions.ReducedDM — Method

reducedDM(state, reducingIndices)

Reduce the density matrix |state⟩⟨state| by tracing over the indices not specified in reducingIndices.

Examples

julia> state = Dict(Bool.([1, 0]) => 0.5, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 0.5
  [0, 1] => -0.5

julia> reducedDM(state, [1])
2×2 Matrix{Float64}:
 0.5  0.0
 0.0  0.5

fermions.VonNEntropy — Method

VonNEntropy(state, reducingIndices)

Calculate entanglement entropy of the subsystem defined by reducingIndices.

Examples

julia> state = Dict(Bool.([1, 0]) => 0.5, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 0.5
  [0, 1] => -0.5

julia> VonNEntropy(state, [1])
0.6931471805599453

fermions.MutInfo — Method

MutInfo(state, reducingIndices)

Calculate mutual information between the subsystems defined by the tuple reducingIndices.

Examples

julia> state = Dict(Bool.([1, 0, 1]) => 0.5, Bool.([0, 1, 0]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [0, 1, 0] => -0.5
  [1, 0, 1] => 0.5

julia> MutInfo(state, ([1], [2]))
0.6931471805599453

fermions.TripartiteInfo — Method

TripartiteInfo(state, reducingIndices)

Calculate tripartite information between the subsystems defined by the 3-tuple reducingIndices.

Examples

julia> state = Dict(Bool.([1, 0, 1, 0]) => 0.5, Bool.([0, 1, 0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0, 1, 0] => 0.5
  [0, 1, 0, 1] => -0.5

julia> TripartiteInfo(state, ([1], [2], [3]))
0.6931471805599453

fermions.ThermalAverage — Method

ThermalAverage(eigenStates, eigenVals, operator, invTemp)

Calculate the canonical ensemble average of operator at the given inverse temperature for the given spectrum.

Examples

julia> basis = BasisStates(2)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0] => 1.0)
 Dict([0, 1] => 1.0)
 Dict([1, 0] => 1.0)
 Dict([1, 1] => 1.0)

julia> eigenStates = BasisStates(2)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0] => 1.0)
 Dict([0, 1] => 1.0)
 Dict([1, 0] => 1.0)
 Dict([1, 1] => 1.0)

julia> eigenVals = rand(4)
4-element Vector{Float64}:
 0.018726877619070992
 0.43796674226812815
 0.7905965730299785
 0.97571966188274

julia> operator = [("n", [1], 1.0)]
1-element Vector{Tuple{String, Vector{Int64}, Float64}}:
 ("n", [1], 1.0)

julia> ThermalAverage(basis, eigenVals, operator, 1.0)
0.33797199716422116

fermions.SpecFunc — Method

SpecFunc(eigVals, eigVecs, probe, probeDiag, freqArray, broadening)

Calculate the spectral function for the excitations defined by probe and probeDiag.

Examples

julia> eigenVals = [0., 1., 1.];

julia> eigenStates = Dict{BitVector, Float64}[Dict([1, 0] => 1.0, [0, 1] => 1.0), Dict([1, 1] => 1.0), Dict([0, 0] => 1.0)]
3-element Vector{Dict{BitVector, Float64}}:
 Dict([1, 0] => 1.0, [0, 1] => 1.0)
 Dict([1, 1] => 1.0)
 Dict([0, 0] => 1.0)

julia> probe = [("-", [1], 1.0)];

julia> probeDag = [("+", [1], 1.0)];

julia> freqArray = collect(range(-2, stop=2, length=10));

julia> SpecFunc(eigenVals, eigenStates, probe, probeDag, freqArray, 1e-2)
10-element Vector{Float64}:
 0.011110098865559272
 0.03392067328129922
 0.8057354340607891
 0.09351894323911442
 0.023221646866030225
 0.023221646866030225
 0.09351894323911442
 0.8057354340607891
 0.03392067328129922
 0.011110098865559272

fermions.SpecFunc — Method

SpecFunc(eigVals, eigVecs, probe, probeDiag, freqArray, broadening, symmetries)

Extends SpecFunc() by making use of symmetries.

Examples

julia> SpecFunc(eigenVals, eigenStates, probe, probeDag, freqArray, 1e-2, ['N'])