Base - Working with States and Operators

fermions.BasisStates — Method

BasisStates(numLevels, totOccReq, magzReq, localCriteria)

Creates a set of basis states for the numLevels Fock states, having the required total occupancy and magnetization, and subject to any criteria imposed locally.

Examples

julia> BasisStates(2, [1], [0], x->true)
Dict{BitVector, Float64}[]
julia> BasisStates(3, [1], [-1], x->true)
1-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 1, 0] => 1.0)
julia> BasisStates(4, [2], [0], x->true)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0, 1, 1] => 1.0)
 Dict([0, 1, 1, 0] => 1.0)
 Dict([1, 0, 0, 1] => 1.0)
 Dict([1, 1, 0, 0] => 1.0)
julia> BasisStates(4, [2], [0], x->sum(x[1:2])==1)
2-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 1, 1, 0] => 1.0)
 Dict([1, 0, 0, 1] => 1.0)

fermions.TransformBit — Method

TransformBit(qubit, operator)

Apply the single qubit operator ('n', 'h', '+' or '-') on a single fock state.

Examples

julia> TransformBit(Bool(0), '+')
(1, 1)

julia> TransformBit(Bool(1), 'h')
(1, 0)

julia> TransformBit(Bool(1), '-')
(0, 1)

fermions.ApplyOperatorChunk — Method

ApplyOperatorChunk(opType, opMembers, opStrength, incomingState; tolerance)

Apply a single tensor product operator chunk (for eg., c^†1 c2 or n1 c^†3 c_4) on a general state and return the new state.

Examples

julia> state = Dict(Bool.([1, 0]) => 1.0, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 1.0
  [0, 1] => -0.5

julia> opType, opMembers, opStrength = ("+-", [1, 2], 0.1)
("+-", [1, 2], 0.1)

julia> ApplyOperatorChunk(opType, opMembers, opStrength, state)
Dict{BitVector, Float64} with 1 entry:
  [1, 0] => -0.05

fermions.ApplyOperator — Method

ApplyOperator(operator, incomingState; tolerance)

Extends ApplyOperatorChunk() by applying a more general operator (consisting of multiple operator chunks) on a general state.

Examples

julia> state = Dict(Bool.([1, 0]) => 1.0, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 1.0
  [0, 1] => -0.5

julia> operator = [("+-", [1, 2], 0.1), ("nh", [2, 1], 1.0)]
2-element Vector{Tuple{String, Vector{Int64}, Float64}}:
 ("+-", [1, 2], 0.1)
 ("nh", [2, 1], 1.0)

julia> ApplyOperator(operator, state)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => -0.05
  [0, 1] => -0.5

fermions.OperatorMatrix — Method

OperatorMatrix(basisStates, operator)

Return the matrix representation of the operator in the given basis.

Examples

julia> basis = BasisStates(2)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0] => 1.0)
 Dict([0, 1] => 1.0)
 Dict([1, 0] => 1.0)
 Dict([1, 1] => 1.0)

julia> operator = [("+-", [1, 2], 0.5), ("n", [2], -1.0)]
2-element Vector{Tuple{String, Vector{Int64}, Float64}}:
 ("+-", [1, 2], 0.5)
 ("n", [2], -1.0)

julia> OperatorMatrix(basis, operator)
4×4 Matrix{Float64}:
 0.0   0.0  0.0   0.0
 0.0  -1.0  0.0   0.0
 0.0   0.5  0.0   0.0
 0.0   0.0  0.0  -1.0

fermions.StateOverlap — Method

StateOverlap(state1, state2)

Compute the inner product ⟨state1|state2⟩.

Examples

julia> state1 = Dict(Bool.([1, 0]) => 1.0, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 1.0
  [0, 1] => -0.5

julia> state2 = Dict(Bool.([1, 1]) => 0.5, Bool.([0, 1]) => 0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 1] => 0.5
  [0, 1] => 0.5

julia> StateOverlap(state1, state2)
-0.25