Base - Working with States and Operators

Fermions.BasisStates — Method

BasisStates(numLevels, totOccReq, magzReq, localCriteria)

Creates a set of basis states for the numLevels Fock states, having the required total occupancy and magnetization, and subject to any criteria imposed locally.

Examples

julia> BasisStates(2, [1], [0], x->true)
Dict{BitVector, Float64}[]
julia> BasisStates(3, [1], [-1], x->true)
1-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 1, 0] => 1.0)
julia> BasisStates(4, [2], [0], x->true)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0, 1, 1] => 1.0)
 Dict([0, 1, 1, 0] => 1.0)
 Dict([1, 0, 0, 1] => 1.0)
 Dict([1, 1, 0, 0] => 1.0)
julia> BasisStates(4, [2], [0], x->sum(x[1:2])==1)
2-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 1, 1, 0] => 1.0)
 Dict([1, 0, 0, 1] => 1.0)

Fermions.BasisStates — Method

BasisStates(numLevels; totOccReq, magzReq, localCriteria)

Extends BasisStates() by making the last three arguments optional. When skipped, these arguments are filled with all possible values and then passed to BasisStates().

Examples

julia> BasisStates(2)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0] => 1.0)
 Dict([0, 1] => 1.0)
 Dict([1, 0] => 1.0)
 Dict([1, 1] => 1.0)

julia> BasisStates(2; totOccReq=1)
2-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 1] => 1.0)
 Dict([1, 0] => 1.0)

julia> BasisStates(2; magzReq=0)
2-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0] => 1.0)
 Dict([1, 1] => 1.0)

Fermions.BasisStates1p — Method

BasisStates(numLevels)

Specialises BasisStates() by restricting to the 1-particle case; returns all 1-particle basis states for the given number of levels. Equivalent to BasisStates(numLevels; totOccReq=1), but much much faster.

Examples

julia> BasisStates1p(10)
10-element Vector{Dict{BitVector, Float64}}:
 Dict([1, 0, 0, 0, 0, 0, 0, 0, 0, 0] => 1.0)
 Dict([0, 1, 0, 0, 0, 0, 0, 0, 0, 0] => 1.0)
 Dict([0, 0, 1, 0, 0, 0, 0, 0, 0, 0] => 1.0)
 Dict([0, 0, 0, 1, 0, 0, 0, 0, 0, 0] => 1.0)
 Dict([0, 0, 0, 0, 1, 0, 0, 0, 0, 0] => 1.0)
 Dict([0, 0, 0, 0, 0, 1, 0, 0, 0, 0] => 1.0)
 Dict([0, 0, 0, 0, 0, 0, 1, 0, 0, 0] => 1.0)
 Dict([0, 0, 0, 0, 0, 0, 0, 1, 0, 0] => 1.0)
 Dict([0, 0, 0, 0, 0, 0, 0, 0, 1, 0] => 1.0)
 Dict([0, 0, 0, 0, 0, 0, 0, 0, 0, 1] => 1.0)

Fermions.TransformBit — Method

TransformBit(qubit, operator)

Apply the single qubit operator ('n', 'h', '+' or '-') on a single fock state.

Examples

julia> TransformBit(Bool(0), '+')
(1, 1)

julia> TransformBit(Bool(1), 'h')
(1, 0)

julia> TransformBit(Bool(1), '-')
(0, 1)

Fermions.ApplyOperatorChunk — Method

ApplyOperatorChunk(opType, opMembers, opStrength, incomingState; tolerance)

Apply a single tensor product operator chunk (for eg., c^†1 c2 or n1 c^†3 c_4) on a general state and return the new state.

Examples

julia> state = Dict(Bool.([1, 0]) => 1.0, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 1.0
  [0, 1] => -0.5

julia> opType, opMembers, opStrength = ("+-", [1, 2], 0.1)
("+-", [1, 2], 0.1)

julia> ApplyOperatorChunk(opType, opMembers, opStrength, state)
Dict{BitVector, Float64} with 1 entry:
  [1, 0] => -0.05

Fermions.ApplyOperator — Method

ApplyOperator(operator, incomingState; tolerance)

Extends ApplyOperatorChunk() by applying a more general operator (consisting of multiple operator chunks) on a general state.

Examples

julia> state = Dict(Bool.([1, 0]) => 1.0, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 1.0
  [0, 1] => -0.5

julia> operator = [("+-", [1, 2], 0.1), ("nh", [2, 1], 1.0)]
2-element Vector{Tuple{String, Vector{Int64}, Float64}}:
 ("+-", [1, 2], 0.1)
 ("nh", [2, 1], 1.0)

julia> ApplyOperator(operator, state)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => -0.05
  [0, 1] => -0.5

Fermions.OperatorMatrix — Method

OperatorMatrix(basisStates, operator)

Return the matrix representation of the operator in the given basis.

Examples

julia> basis = BasisStates(2)
4-element Vector{Dict{BitVector, Float64}}:
 Dict([0, 0] => 1.0)
 Dict([0, 1] => 1.0)
 Dict([1, 0] => 1.0)
 Dict([1, 1] => 1.0)

julia> operator = [("+-", [1, 2], 0.5), ("n", [2], -1.0)]
2-element Vector{Tuple{String, Vector{Int64}, Float64}}:
 ("+-", [1, 2], 0.5)
 ("n", [2], -1.0)

julia> OperatorMatrix(basis, operator)
4×4 Matrix{Float64}:
 0.0   0.0  0.0   0.0
 0.0  -1.0  0.0   0.0
 0.0   0.5  0.0   0.0
 0.0   0.0  0.0  -1.0

Fermions.StateOverlap — Method

StateOverlap(state1, state2)

Compute the inner product ⟨state1|state2⟩.

Examples

julia> state1 = Dict(Bool.([1, 0]) => 1.0, Bool.([0, 1]) => -0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 0] => 1.0
  [0, 1] => -0.5

julia> state2 = Dict(Bool.([1, 1]) => 0.5, Bool.([0, 1]) => 0.5)
Dict{BitVector, Float64} with 2 entries:
  [1, 1] => 0.5
  [0, 1] => 0.5

julia> StateOverlap(state1, state2)
-0.25

Fermions.ExpandIntoBasis — Method

ExpandIntoBasis(state, basisStates)

Returns the basis decomposition of the provided state. That is, if the basis is {bi} and the state is {b1=>c1, b2=>c2, ...}, the function returns [c1, c_2, ...].

Fermions.GetSector — Method

GetSector(state, symmetries)

Returns the symmetry sector of the provided state. If symmetries only has 'N'('Z'), returns the total occupancy (total magnetization).

Fermions.RoundTo — Method

RoundTo(val, tolerance)

Round the provided float to the specified tolerance.

Examples

julia> fermions.roundTo(1.122323, 1e-3)
1.122

Fermions.PermuteSites — Method

PermuteSites(state, permutation)

Given a basis state |n{ν1} n{ν2} n{ν3} … ⟩ corresponding to a particular ordering choice [ν1, ν_2, …], this function permutes the basis into P(B) = [P(ν1), P(ν2),P(ν3) … ], and returns the basis state written in this new basis, accounting for any fermions signs arising from index permutations.

Examples

julia> state = Bool.([1, 1])
2-element BitVector:
 1
 1

julia> PermuteSites(state, [2, 1])
(Bool[1, 1], -1)

Fermions.PermuteSites — Method

PermuteSites(state, permutation)

Given a state that is written in a particular basis B = {|n{ν1} n{ν2} n{ν3} … ⟩}, this function permutes the basis into P(B) = {|n{P(ν1)} n{P(ν2)} n{P(ν3)} … ⟩} and returns the state written in this new basis, accounting for any fermions signs arising from index permutations.

Examples

julia> state
Dict{BitVector, Float64} with 2 entries:
  [1, 1] => 0.5
  [0, 1] => 0.3

julia> PermuteSites(state, [2, 1])
Dict{BitVector, Float64} with 2 entries:
  [1, 1] => -0.5
  [1, 0] => 0.3

Fermions.Dagger — Method

Dagger(operator, members)

Native method to calculate hermitian conjugate of operator.

Examples

jldoctest julia> Dagger("+--+", [1,4,3,2]) ("-++-", [2, 3, 4, 1])

Fermions.Dagger — Method

Dagger(operator, members)

Native method to calculate hermitian conjugate of operator.

Examples

```jldoctest julia> operator = [("++", [1, 2], 1.), ("+n", [3, 4], 2.)] 2-element Vector{Tuple{String, Vector{Int64}, Float64}}: ("++", [1, 2], 1.0) ("+n", [3, 4], 2.0)

julia> Dagger(operator) 2-element Vector{Tuple{String, Vector{Int64}, Float64}}: ("–", [2, 1], 1.0) ("n-", [4, 3], 2.0) ```